finished battery collection proto
- added battery object
- collision detection for picking up battery
- ui number to show how many batteries collected

+++ Assets/Directory/Scripts/Player_SO.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Player_SO : ScriptableObject
 {
     public float progression;
-    public float batteries;
+    public int batteries;
